+++ test/view_component/integration_test.rb
@@ -386,4 +386,68 @@ class IntegrationTest < ActionDispatch::IntegrationTest
       assert_includes response.body, "bar"
     end
   end
+
+  test "renders the inline component preview examples with default behaviour and with their own templates" do
+    get "/rails/view_components/inline_component/default"
+    assert_select "input" do
+      assert_select "[name=?]", "name"
+    end
+
+    get "/rails/view_components/inline_component/inside_form"
+    assert_select "form" do
+      assert_select "p", "Inside Form"
+      assert_select "input[name=?]", "name"
+    end
+
+    get "/rails/view_components/inline_component/outside_form"
+    assert_select "div" do
+      assert_select "p", "Outside Form"
+      assert_select "input[name=?]", "name"
+    end
+  end
+
+  test "renders the preview example with its own template and a layout" do
+    get "/rails/view_components/my_component/inside_banner"
+    assert_includes response.body, "ViewComponent - Admin - Test"
+    assert_select ".banner" do
+      assert_select("div", "hello,world!")
+    end
+  end
+
+  test "renders an inline component preview using URL params and a template" do
+    get "/rails/view_components/inline_component/with_params?form_title=This is a test form"
+    assert_select "form" do
+      assert_select "p", "This is a test form"
+      assert_select "input[name=?]", "name"
+    end
+  end
+
+  test "renders the inline component using a non standard-located template" do
+    get "/rails/view_components/inline_component/with_non_standard_template"
+    assert_select "h1", "This is not a standard place to have a preview template"
+    assert_select "input[name=?]", "name"
+  end
+
+  test "renders an inline component preview using a HAML template" do
+    get "/rails/view_components/inline_component/with_haml"
+    assert_select "h1", "Some HAML here"
+    assert_select "input[name=?]", "name"
+  end
+
+  test "raises an error if the template is not present and the render_with_template method is used in the example" do
+    error = assert_raises ViewComponent::PreviewTemplateError do
+      get "/rails/view_components/inline_component/without_template"
+    end
+    assert_match /preview template for example without_template does not exist/, error.message
+  end
+
+  test "renders a preview template using HAML, params from URL, custom template and locals" do
+    get "/rails/view_components/inline_component/with_several_options?form_title=Title from params"
+
+    assert_select "form" do
+      assert_select "h1", "Title from params"
+      assert_select "input[name=?]", "name"
+      assert_select "input[value=?]", "Send this form!"
+    end
+  end
 end